# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Trade Finder API
  version: "2.0"
  description: |
    OpenAPI specification for the AI Trade Finder project.

    Notes:
    - Authentication is not implemented (all endpoints are open).
    - Some controllers return empty bodies on error (e.g., 500 with no JSON).
    - Several fields are intentionally flexible objects (Map<String,Object>); those are modeled as `type: object` with `additionalProperties: true`.
servers:
  - url: http://localhost:8082
    description: Local default
...
paths:
...
  /api/v2/schedulers:
    post:
      tags: [Schedulers v2]
      summary: Create a scheduler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulerConfig'
            examples:
              fixed_rate:
                value:
                  name: custom-analysis-es
                  description: Custom ES analysis every 10 minutes
                  type: CUSTOM
                  enabled: true
                  scheduleExpression: '600000'
                  scheduleType: FIXED_RATE
                  parameters:
                    symbol: ES
                    timeframes: [5m, 15m]
                    minConfidence: 70
                  handlerBean: aiTradeFinderService
                  handlerMethod: findTrades
                  priority: 5
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerConfig'
        '409':
          description: Name conflict (often empty body)
        '500':
          description: Internal error (often empty body)
    get:
      tags: [Schedulers v2]
      summary: List schedulers
      responses:
        '200':
          description: Schedulers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchedulerConfig'
  /api/v2/schedulers/{name}:
    get:
      tags: [Schedulers v2]
      summary: Get scheduler by name
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Scheduler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerConfig'
        '404':
          description: Not found
    put:
      tags: [Schedulers v2]
      summary: Update scheduler by name
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulerConfig'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerConfig'
        '500':
          description: Internal error (often empty body, includes not-found)
    delete:
      tags: [Schedulers v2]
      summary: Delete scheduler by name
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
        '500':
          description: Internal error (often empty body, includes not-found)
  /api/v2/schedulers/{name}/enable:
    post:
      tags: [Schedulers v2]
      summary: Enable scheduler
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated scheduler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerConfig'
        '500':
          description: Internal error (often empty body)
  /api/v2/schedulers/{name}/disable:
    post:
      tags: [Schedulers v2]
      summary: Disable scheduler
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated scheduler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerConfig'
        '500':
          description: Internal error (often empty body)
...
  /api/v1/trade-finder/trades/{symbol}:
    get:
      tags: [Trade Finder v1]
      summary: Get trades by symbol
      parameters:
        - in: path
          name: symbol
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IdentifiedTrade'
  /api/v1/trade-finder/trades/{symbol}/status/{status}:
    get:
      tags: [Trade Finder v1]
      summary: Get trades by symbol and status
      parameters:
        - in: path
          name: symbol
          required: true
          schema: { type: string }
        - in: path
          name: status
          required: true
          schema:
            type: string
            description: Trade status string
            enum: [IDENTIFIED, ALERTED, EXPIRED, CANCELLED, TAKEN, INVALIDATED]
      responses:
        '200':
          description: Trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IdentifiedTrade'
  /api/v1/trade-finder/trades/recent:
    get:
      tags: [Trade Finder v1]
      summary: Get recent trades across all symbols
      parameters:
        - in: query
          name: hours
          required: false
          schema:
            type: integer
            default: 24
            minimum: 1
            maximum: 168
      responses:
        '200':
          description: Trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IdentifiedTrade'
  /api/v1/trade-finder/statistics:
    get:
      tags: [Trade Finder v1]
      summary: Get trade statistics
      parameters:
        - in: query
          name: hours
          required: false
          schema:
            type: integer
            default: 24
            minimum: 1
            maximum: 168
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTradeStatistics'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          description: Response status
          enum: [success, error, warning]
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        data:
          nullable: true
          description: Payload (varies by endpoint)
        metadata:
          nullable: true
          description: Optional metadata (varies by endpoint)
    ApiResponseTradeStatistics:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TradeStatisticsDTO'
...
    TradeStatisticsDTO:
      type: object
      properties:
        periodHours: { type: integer, format: int32 }
        totalTrades: { type: integer, format: int64 }
        activeTrades: { type: integer, format: int64 }
        expiredTrades: { type: integer, format: int64 }
        alertedTrades: { type: integer, format: int64 }
        averageConfidence: { type: number, format: double }
        bySymbol:
          type: object
          additionalProperties: { type: integer, format: int64 }
        byDirection:
          type: object
          additionalProperties: { type: integer, format: int64 }
        timestamp:
          type: string
          format: date-time
...
    SchedulerConfig:
      type: object
      required: [name]
      properties:
        id: { type: string, nullable: true }
        name: { type: string }
        description: { type: string, nullable: true }
        type:
          type: string
          description: TRADE_FINDER | EXPIRATION | STATISTICS | CUSTOM (stored as string)
        enabled: { type: boolean }
        scheduleExpression: { type: string, nullable: true }
        scheduleType:
          type: string
          enum: [CRON, FIXED_RATE, FIXED_DELAY]
        parameters:
          type: object
          nullable: true
          additionalProperties: true
        handlerBean: { type: string, nullable: true }
        handlerMethod: { type: string, nullable: true }
        priority: { type: integer, format: int32, nullable: true }
        lastExecution: { type: string, format: date-time, nullable: true }
        nextExecution: { type: string, format: date-time, nullable: true }
        executionCount: { type: integer, format: int64, nullable: true }
        failureCount: { type: integer, format: int64, nullable: true }
        lastStatus: { type: string, nullable: true }
        lastError: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        createdBy: { type: string, nullable: true }
...
    IdentifiedTrade:
      type: object
      properties:
        id: { type: string }
        symbol: { type: string }
        direction:
          type: string
          enum: [LONG, SHORT]
        identifiedAt: { type: string, format: date-time, nullable: true }
        confidence: { type: integer, format: int32, nullable: true, minimum: 0, maximum: 100 }
        status:
          type: string
          enum: [IDENTIFIED, ALERTED, EXPIRED, CANCELLED, TAKEN, INVALIDATED]
        entryZoneType:
          type: string
          enum: [FVG_CE, IFVG, OB, BREAKER, MITIGATION]
          nullable: true
        entryZone: { type: string, nullable: true }
        entryPrice: { type: number, format: double, nullable: true }
        stopPlacement: { type: string, nullable: true }
        targets:
          type: array
          items: { type: string }
          nullable: true
        rrHint: { type: string, nullable: true }
        narrative: { type: string, nullable: true }
        triggerConditions:
          type: array
          items: { type: string }
          nullable: true
        invalidations:
          type: array
          items: { type: string }
          nullable: true
        sessionLabel: { type: string, nullable: true }
        timeframe: { type: string, nullable: true }
        dedupeKey: { type: string, nullable: true }
        alertSent: { type: boolean }
        alertSentAt: { type: string, format: date-time, nullable: true }
        alertType:
          type: string
          nullable: true
          enum: [CALL_SMS_TELEGRAM, SMS_TELEGRAM, LOG_ONLY]
        createdAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        aiRequestId: { type: string, nullable: true }
        aiFullResponse: { type: string, nullable: true }
        keyLevelsSnapshot: { type: string, nullable: true }
...